+++ e2e/tests/core-sdk.test.ts
@@ -719,7 +719,7 @@ async function raiseDispute(
   expect(exchangeAfterDispute.completedDate).toBeNull();
   expect(exchangeAfterDispute.finalizedDate).toBeNull();
 
-  // exchange state is now DISPUTED
+  expect(exchangeAfterDispute.disputedDate).toBeTruthy();
   expect(exchangeAfterDispute.disputed).toBeTruthy();
 }
 
@@ -773,6 +773,7 @@ async function checkDisputeRetracted(exchangeId: string, coreSDK: CoreSDK) {
 
   // dispute finalizedDate is correct
   expect(dispute.finalizedDate).toBeTruthy();
+  expect(dispute.retractedDate).toBeTruthy();
 }
 
 async function resolveDispute(
@@ -787,9 +788,9 @@ async function resolveDispute(
       r: sigR,
       s: sigS,
       v: sigV
-    } = await signerSDK.signMutualAgreement({
-      exchangeId: exchangeId,
-      buyerPercent: buyerPercent
+    } = await signerSDK.signDisputeResolutionProposal({
+      exchangeId,
+      buyerPercent
     });
 
     // send the Resolve transaction from buyer
@@ -831,6 +832,7 @@ async function checkDisputeResolved(
 
   // dispute finalizedDate is correct
   expect(dispute.finalizedDate).toBeTruthy();
+  expect(dispute.resolvedDate).toBeTruthy();
 }
 
 async function escalateDispute(exchangeId: string, buyerCoreSDK: CoreSDK) {
@@ -910,6 +912,7 @@ async function checkDisputeDecided(
 
   // dispute finalizedDate is filled
   expect(dispute.finalizedDate).toBeTruthy();
+  expect(dispute.decidedDate).toBeTruthy();
 
   // dispute buyerPercent is filled
   expect(dispute.buyerPercent).toEqual(buyerPercent);
@@ -966,6 +969,7 @@ async function checkDisputeRefused(exchangeId: string, coreSDK: CoreSDK) {
 
   // dispute finalizedDate is filled
   expect(dispute.finalizedDate).toBeTruthy();
+  expect(dispute.refusedDate).toBeTruthy();
 
   // dispute buyerPercent is filled
   expect(dispute.buyerPercent).toEqual("0");