+++ packages/common/src/types/configs.ts
@@ -6,10 +6,14 @@ export type ContractAddresses = {
 export type ProtocolConfig = {
   envName: string;
   chainId: number;
+  nativeCoin?: {
+    symbol: string;
+    name: string;
+    decimals: string;
+  };
   subgraphUrl: string;
   theGraphIpfsUrl?: string;
   jsonRpcUrl: string;
   ipfsMetadataUrl: string;
-  widgetsUrl: string;
   contracts: ContractAddresses;
 };